requirements.txt
google-generativeai==0.8.3
beautifulsoup4>=4.12
youtube-transcript-api>=0.6
numpy>=1.24
pandas>=2.0
chromadb>=0.5
streamlit>=1.35
tqdm
requests
